% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Generate a map of all incidents in a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer from 0 through 51, representing each state and the District of Columbia}

\item{year}{An integer representing a year that is 2013, 2014, or 2015}
}
\value{
A map of the state with each dot representing an incident.  If there are no accidents in that state
in the given year, the function returns a message as such.  If a number is input that does not correspond
to a state, the function returns a message that the provided state number is invalid
}
\description{
This function uses the \code{maps} package to display dots for each incident on a map of a given state over
a given year.  It takes as inputs the state, represented by its number with all states ordered alphabetically,
and year.  The \code{filter} function from the \code{dplyr} package is used to consolidate the data
}
\examples{
fars_map_state(5, 2013)
fars_map_state(51, 2014)

}
