% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Import FARS data as data frame}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A bz2 file or a string pointing to such a file}
}
\value{
This function returns a data frame of all of the FARS data from the bz2 file
}
\examples{
fars_read("accident_2013.csv.bz2")
fars_read("notfound")

}
