% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Generate a table identifying the month of each datapoint}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer representing a year that is 2013, 2014, or 2015}
}
\value{
A table with the same number of rows as the data file associated with the year being evaluated,
   identifying the month associated with each datapoint
}
\description{
This function uses the \code{mutate} and \code{select} functions from the \code{dplyr} package to strip
the full data frame associated with a given year, leaving only the rows associated with months and the year.
This provides a matrix of the number of rows of data from a given year, and columns of months, expressed
as an integer from 1 to 12, and the year.  If a year for which data is not available is input, function
alerts user that it is an invalid year.
}
\details{
makes use of the \code{\link{make_filename}} and \code{\link{fars_read}} functions by generating a
   filename based on the year provided, and then generating a data frame from the data associated with that
   year's datafile
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(2014)
fars_read_years(2015)
}

}
