% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Group the number of incidents per month in a given year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer representing a year that is 2013, 2014, or 2015}
}
\value{
A table with the total number of incidents in each individual month of the given year
}
\description{
This function returns the total number of incidents per month in a given year. The only input is the
year the user wishes to evaluate, and the \code{dplyr} package's \code{bind_rows}, \code{group_by}, and
\code{summarize} functions and the \code{tidyr} package's \code{spread} function are all used to aggregate
the individual datapoints into a table of months and incidents per month
}
\details{
makes use of the \code{\link{fars_read_years}} function, which uses the \code{\link{make_filename}}
    and \code{\link{fars_read}} functions, to first make a matrix of the same number of rows as the primary
    datafile, then to add the relevant incident data to each row, and then finally to summarize each month
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(2014)
fars_summarize_years(2015)

}
