% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate filename based on specified year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer representing a year that is 2013, 2014, or 2015}
}
\value{
This function returns a character string of the full bz2 filename
}
\description{
This function generates a filename that can be used in the fars_read function. It takes as an input
a specific year, and provides as output the full filename as a character string.
}
\examples{
make_filename(2013)
make_filename(2014)
make_filename(2015)

}
